Database completely updates and fuzzy kinda works

+++ app/src/main/java/com/example/journalentry/database/Entity.kt
@@ -1,4 +1,4 @@
 package com.example.journalentry.database
 
-data class Entity(val key: String, val value: String, val sentenceId: Int, val type: String, val score: Int) {
+data class Entity(val key: String, val value: String, val sentenceId: Int, val type: String, val score: Float) {
 }
